requirements.txt
comtypes==1.2.0
packaging==23.1
Pillow==10.0.0
pypiwin32==223
pytesseract==0.3.10
pyttsx3==2.90
pywin32==306
